Attempt to fix signed-vs.-unsigned issues revealed by -Wextra.
The theme in this patch is changing variables that are semantically data
lengths from int to size_t.  This cleanly resolves many sign-compare bugs.
It requires some printfs to change from %d to %zd formats.

No logic changes. Associated code has been eyeball-checked for subtractions that
might induce modular wraparound, but if a future core dump produces a bisection
that lands here that will be the thing to suspect.

+++ include/ntp_refclock.h
@@ -235,7 +235,7 @@ extern 	void	refclock_process_offset(struct refclockproc *, l_fp,
 					l_fp, double);
 extern	void	refclock_report	(struct peer *, int);
 extern	int	refclock_gtlin	(struct recvbuf *, char *, int, l_fp *);
-extern	int	refclock_gtraw	(struct recvbuf *, char *, int, l_fp *);
+extern	size_t	refclock_gtraw	(struct recvbuf *, char *, size_t, l_fp *);
 extern	bool	indicate_refclock_packet(struct refclockio *,
 					 struct recvbuf *);
 extern	void	process_refclock_packet(struct recvbuf *);
